requirements.txt
aniso8601==7.0.0
astroid==2.3.3
boto3==1.10.37
botocore==1.13.37
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
django-graphql-jwt==0.2.2
django-graphql-social-auth==0.1.4
django-hitcount==1.3.1
django-sslserver==0.21
django-storages==1.8
django-summernote==0.8.11.4
docutils==0.15.2
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.7.1
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.2.1
promise==2.2.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
raven==6.10.0
requests==2.22.0
requests-oauthlib==1.3.0
Rx==1.6.1
s3transfer==0.2.1
singledispatch==3.4.0.3
six==1.12.0
slacker==0.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
whitenoise==3.3.1
wrapt==1.11.2
